add download_file scaffold; move downloader to common
